+++ flink-contrib/flink-statebackend-rocksdb/src/test/java/org/apache/flink/contrib/streaming/state/benchmark/RocksDBListStatePerformanceTest.java
@@ -19,7 +19,6 @@
 package org.apache.flink.contrib.streaming.state.benchmark;
 
 import org.apache.flink.contrib.streaming.state.RocksDBKeyedStateBackend;
-import org.apache.flink.contrib.streaming.state.util.MergeUtils;
 import org.apache.flink.testutils.junit.RetryOnFailure;
 import org.apache.flink.testutils.junit.RetryRule;
 import org.apache.flink.util.TestLogger;
@@ -62,6 +61,8 @@ import java.util.List;
  *
  */
 public class RocksDBListStatePerformanceTest extends TestLogger {
+
+	private static final byte DELIMITER = ',';
 
 	@Rule
 	public final TemporaryFolder tmp = new TemporaryFolder();
@@ -121,7 +122,7 @@ public class RocksDBListStatePerformanceTest extends TestLogger {
 			for (int i = 0; i < num; i++) {
 				list.add(valueBytes);
 			}
-			byte[] premerged = MergeUtils.merge(list);
+			byte[] premerged = merge(list);
 
 			log.info("begin update");
 
@@ -131,5 +132,37 @@ public class RocksDBListStatePerformanceTest extends TestLogger {
 
 			log.info("end update - duration: {} ns", (endInsert2 - beginInsert2));
 		}
+	}
+
+	/**
+	 * Merge operands into a single value that can be put directly into RocksDB.
+	 */
+	public static byte[] merge(List<byte[]> operands) {
+		if (operands == null || operands.size() == 0) {
+			return null;
+		}
+
+		if (operands.size() == 1) {
+			return operands.get(0);
+		}
+
+		int numBytes = 0;
+		for (byte[] arr : operands) {
+			numBytes += arr.length + 1;
+		}
+		numBytes--;
+
+		byte[] result = new byte[numBytes];
+
+		System.arraycopy(operands.get(0), 0, result, 0, operands.get(0).length);
+
+		for (int i = 1, arrIndex = operands.get(0).length; i < operands.size(); i++) {
+			result[arrIndex] = DELIMITER;
+			arrIndex += 1;
+			System.arraycopy(operands.get(i), 0, result, arrIndex, operands.get(i).length);
+			arrIndex += operands.get(i).length;
+		}
+
+		return result;
 	}
 }
